igsc: oprom: fix type size comparison issue in oprom_parser.c
Fixed a type size comparison issue in debug_print_device_type_ext()
function, when comparing len and extension_length in the for loop.

+++ lib/oprom_parser.c
@@ -91,7 +91,7 @@ struct igsc_oprom_image {
 static void debug_print_device_type_ext(struct mft_oprom_device_type_ext *ext)
 {
     struct oprom_subsystem_device_id *dev = &ext->device_ids[0];
-    uint32_t len = sizeof(struct mft_ext_header_with_data);
+    size_t len = sizeof(struct mft_ext_header_with_data);
 
     gsc_debug("type %u len %u\n", ext->extension_type, ext->extension_length);
     for (; len < ext->extension_length; len += sizeof(*dev))
